third_party/WebKit/LayoutTests/fast/forms/form-dirname-attribute.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../resources/js-test.js"> </script>
</head>
<body>
<p>"dirname" attribute</p>
<pre id="console"></pre>
<script>
var input = document.createElement('input');
input.setAttribute('dirName', "Hello");
shouldBeEqualToString('input.dirName', "Hello");

var textArea = document.createElement('textarea');
textArea.setAttribute('dirName', "Hello");
shouldBeEqualToString('textArea.dirName', "Hello");
successfullyParsed = true;
</script>
</body>
</html>
